Fix up prod build for speed on netlify

+++ build_prod.sh
@@ -8,34 +8,39 @@ if [ $INSTALL_WEBPACK ]; then
   echo "##> Installed webpack."
 fi
 
-# prep for netlify cache stuff
-echo "Restoring .cache at $CACHE_DIR/.cache: \n`ls $CACHE_DIR/.cache`\n"
-cp -a $CACHE_DIR/.cache .cache || true
+# detect netlify (or other ci)
+if [ $REPOSITORY_URL ]; then
+  # prep for netlify cache stuff
+  echo "Restoring .cache at $CACHE_DIR/.cache: \n`ls $CACHE_DIR/.cache`\n"
+  cp -a $CACHE_DIR/.cache .cache || true
 
-# get cached elm stuff
-mkdir -p .cache
-cp -a .cache/elm-stuff elm-stuff || true
+  # get cached elm stuff
+  mkdir -p .cache
+  cp -a .cache/elm-stuff elm-stuff || true
 
-# print cache directory
-echo "Cache Directory before build:\n"
-ls -al .cache
-echo ""
+  # print cache directory
+  echo "Cache Directory before build:\n"
+  ls -al .cache
+  echo ""
+fi
 
 # do build
 echo "Building now..."
-webpack "$@"
+time webpack "$@" --progress
 echo "Build Complete"
 
-# save in cache
-echo "Caching elm stuff"
-rm -rf .cache/elm-stuff || true
-cp -a elm-stuff .cache/elm-stuff
+if [ $REPOSITORY_URL ]; then
+  # save in cache
+  echo "Caching elm stuff"
+  rm -rf .cache/elm-stuff || true
+  cp -a elm-stuff .cache/elm-stuff
 
-# print cache directory
-echo "Cache Directory after build: (ls -al .cache)\n"
-ls -al .cache
-echo ""
+  # print cache directory
+  echo "Cache Directory after build: (ls -al .cache)\n"
+  ls -al .cache
+  echo ""
 
-# save netlify cache
-cp -a .cache $CACHE_DIR/.cache
-echo "Saved .cache to $CACHE_DIR/.cache: \n`ls $CACHE_DIR/.cache`\n"
+  # save netlify cache
+  cp -a .cache $CACHE_DIR/.cache
+  echo "Saved .cache to $CACHE_DIR/.cache: \n`ls $CACHE_DIR/.cache`\n"
+fi
